requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
openai==1.58.1
django-password-validators
django-cors-headers
mysqlclient